+++ CHANGELOG.md
@@ -14,3 +14,4 @@
     8.  0.  *   optimizing
     9.  0.  *   only run `require('fs')` when it's really used.
             *   do not check `require('path').isAbsolute(source.path)` anymore.
+    10. 0.  *   update dependency `@ltd/j-orderify` (use `copyWithin` to optimize `ownKeys` `splice` instead of modifying `constructor[Symbol.species]`).